KS-8-3-ingredient-search-without-loop Изменил метод поиска ингредиента с цикла на indexOf.
