+++ src/main/kotlin/lesson_8/task_1.kt
@@ -1,21 +1,4 @@
 package org.example.lesson_8
 
 fun main() {
-    val ingredientArray = arrayOf(
-        "картофель",
-        "помидор",
-        "соль",
-        "перец черный горошек",
-        "базилик",
-        "говядина",
-        "лавровый лист",
-    )
-    println("Введите название ингредиента, который необходимо найти.")
-    val ingredient = readln()
-    val index = ingredientArray.indexOf(ingredient)
-    if (index >= 0) {
-        println("Ингредиент $ingredient в рецепте есть.")
-    } else {
-        println("Такого ингредиента в рецепте нет.")
-    }
 }
